python/agents/economic-research-agent/pyproject.toml
[project]
name = "economic-research"
version = "0.1.0"
description = ""
authors = [
    {name = "Enrique Chan", email = "enriq@google.com"},
    {name = "Casey Justus", email = "caseynjustus@google.com"},
]
dependencies = [
    "census~=0.8.23",
    "db-dtypes~=1.4.2",
    "fastapi>=0.124.1",
    "google-cloud-aiplatform[evaluation]>=1.120.0",
    "google-cloud-secret-manager~=2.23.2",
    "google-cloud-logging~=3.11.4",
    "google-genai>=1.37.0",
    "opentelemetry-exporter-gcp-trace~=1.9.0",
    "pandas~=2.2.3",
    "pylint",
    "streamlit",
    "streamlit-feedback",
    "locust",
    "us~=3.2.0",
    "uvicorn~=0.34.0",
    "plotly~=6.0.0",
    "nbformat>=5.10.4",
    "fredapi>=0.5.2",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "python-dotenv>=1.0.0",
    "google-adk>=1.24.0",
]


requires-python = ">=3.10"


[dependency-groups]
dev = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.23.8",
    "nest-asyncio>=1.6.0",
    "pytest-cov>=4.1.0",
    "agent-starter-pack>=0.29.1",
]

[project.optional-dependencies]
streamlit = [
    "streamlit~=1.42.0",
    "streamlit-extras~=0.4.3",
    "extra-streamlit-components~=0.1.71",
    "streamlit-feedback~=0.1.3",
]
jupyter = [
    "jupyter~=1.0.0",
]
lint = [
    "ruff>=0.4.6",
    "mypy~=1.15.0",
    "codespell~=2.2.0",
    "types-pyyaml~=6.0.12.20240917",
    "types-requests~=2.32.0.20240914",
]

[tool.ruff]
# extends the ruff rules established in the root .toml
extend = "../../../pyproject.toml"

[tool.ruff.lint]
# Ignores function complication rule throughout sample
ignore = ["C901", "PLR0915", "PLR2004", "PLC0415", "PLW2901", "PLR0912"]

[tool.ruff.lint.per-file-ignores]
# ignores unused import rule exclusively for this sample’s __init__.py files
"__init__.py" = ["F401"]

[tool.ruff.lint.isort]
known-first-party = ["economic_research"]

[tool.mypy]
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
no_implicit_optional = true
check_untyped_defs = true
disallow_subclassing_any = true
warn_incomplete_stub = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_unreachable = true
follow_imports = "silent"
ignore_missing_imports = true
explicit_package_bases = true
disable_error_code = ["misc", "no-untyped-call", "no-any-return"]

exclude = [".venv"]

[tool.codespell]
ignore-words-list = "rouge"

skip = "./locust_env/*,uv.lock,.venv,**/*.ipynb"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# This configuration file is used by goo.gle/agent-starter-pack to power remote templating.
# It defines the template's properties and settings.
[tool.agent-starter-pack]
example_question = "Texas vs Ohio for a Data Center"

[tool.agent-starter-pack.settings]
agent_directory = "economic_research"


[tool.pytest.ini_options]
pythonpath = "."

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["economic_research"]
